fix: clear empty unusable worktrees before recreate
After CleanupWorktree on a missing/unusable path, remove empty leftover
directories so CreateWorktree can reuse the managed path, and preserve
populated unregistered leftovers for manual intervention. Real Cleanup
swallows "is not a working tree" without deleting the directory.

Split dirty-adopt recovery tests into helpers, adopt, recovery, real
gateway, and terminal-preserve files under the AGENTS.md growth threshold,
and cover the real-gateway unregistered empty/populated lifecycle.

Generated-By: looper 0.11.0 (runner=fixer, agent=grok-build)
